@@ -4,14 +4,16 @@
 	"description": "Timings for HTTP requests",
 	"main": "source",
 	"engines": {
-		"node": ">=6"
+		"node": ">=8"
 	},
 	"scripts": {
-		"test": "xo && nyc ava && tsd",
+		"test": "xo && nyc ava",
+		"build": "del-cli dist && tsc",
+		"prepublishOnly": "npm run build",
 		"coveralls": "nyc report --reporter=text-lcov | coveralls"
 	},
 	"files": [
-		"source"
+		"dist"
 	],
 	"keywords": [
 		"http",
@@ -30,19 +32,32 @@
 	},
 	"homepage": "https://github.com/szmarczak/http-timer#readme",
 	"xo": {
+		"extends": "xo-typescript",
+		"extensions": [
+			"ts"
+		],
 		"rules": {
-			"unicorn/filename-case": "camelCase"
+			"import/no-unresolved": "off"
 		}
 	},
 	"devDependencies": {
+		"@sindresorhus/tsconfig": "^0.3.0",
+		"@types/node": "^11.13.4",
+		"@typescript-eslint/eslint-plugin": "^1.6.0",
 		"ava": "^0.25.0",
 		"coveralls": "^3.0.2",
+		"defer-to-connect": "^1.0.2",
+		"del-cli": "^1.1.0",
+		"eslint-config-xo-typescript": "^0.9.0",
 		"nyc": "^12.0.2",
 		"p-event": "^2.1.0",
-		"tsd": "^0.7.2",
+		"typescript": "^3.4.3",
 		"xo": "^0.22.0"
 	},
-	"dependencies": {
-		"defer-to-connect": "^1.0.1"
+	"dependencies": {},
+	"nyc": {
+		"extension": [
+			".ts"
+		]
 	}
 }
